avoid server error on delete

+++ src/main/java/com/example/backenddictionnary/backend_dictionary/service/VocabularyService.java
@@ -65,7 +65,8 @@ public class VocabularyService {
 
     public void deleteVocabulary(String word) {
         Vocabulary existingVocabulary = getVocabularyByWord(word);
-
-        VocabularyRepository.deleteById(existingVocabulary.getId());
+        if (existingVocabulary == null) {
+        } else
+            VocabularyRepository.deleteById(existingVocabulary.getId());
     }
 }
